+++ extensions/internet-radio/src/utils.tsx
@@ -1,118 +1,35 @@
 import {
   Action,
   ActionPanel,
-  Color,
   environment,
   Form,
+  getPreferenceValues,
   Icon,
   launchCommand,
   LaunchType,
   LocalStorage,
+  showHUD,
+  showToast,
+  Toast,
   useNavigation,
 } from "@raycast/api";
 import { useState } from "react";
 import { runAppleScript } from "run-applescript";
 import * as defaultStations from "../assets/default-stations.json";
+import { colorMap } from "./genres";
+import { StationData, StationListObject } from "./types";
 
-export const colorMap: { [key: string]: Color } = {
-  "40s": Color.Red,
-  "50s": Color.Orange,
-  "60s": Color.Yellow,
-  "70s": Color.Green,
-  "80s": Color.Blue,
-  rock: Color.Red,
-  pop: Color.Blue,
-  disco: Color.Magenta,
-  alternative: Color.Yellow,
-  indie: Color.Yellow,
-  "classic rock": Color.Red,
-  metal: Color.SecondaryText,
-  "heavy metal": Color.SecondaryText,
-  punk: Color.Purple,
-  ambient: Color.Magenta,
-  "new age": Color.Blue,
-  downtempo: Color.Purple,
-  chillout: Color.Blue,
-  psychill: Color.Purple,
-  soundtracks: Color.Green,
-  classical: Color.Green,
-  symphony: Color.Green,
-  romantic: Color.Red,
-  piano: Color.SecondaryText,
-  jpop: Color.Red,
-  "top 40": Color.Red,
-  oldies: Color.Orange,
-  noise: Color.Magenta,
-  "hard rock": Color.Red,
-  country: Color.Yellow,
-  dubstep: Color.Magenta,
-  electronic: Color.Magenta,
-  "video game music": Color.Green,
-  blues: Color.Blue,
-  jazz: Color.Orange,
-  news: Color.Red,
-  sports: Color.Blue,
-  games: Color.Green,
-  hits: Color.Red,
-  kpop: Color.Magenta,
-  talk: Color.SecondaryText,
-  ASMR: Color.Blue,
-  holiday: Color.Green,
-  folk: Color.Orange,
-  dance: Color.Purple,
-  emo: Color.Purple,
-  opera: Color.Yellow,
-  grunge: Color.Green,
-  goth: Color.Purple,
-  anime: Color.Magenta,
-  acoustic: Color.Orange,
-  gospel: Color.Yellow,
-  guitar: Color.Orange,
-  lullabies: Color.Magenta,
-  "sing along": Color.Magenta,
-  stories: Color.Magenta,
-  "avant-garde": Color.Red,
-  baroque: Color.SecondaryText,
-  choral: Color.Magenta,
-  concerto: Color.Red,
-  expressionist: Color.Red,
-  impressionist: Color.Magenta,
-  orchestral: Color.Orange,
-  novelty: Color.Green,
-  parody: Color.SecondaryText,
-  comedy: Color.Orange,
-  jingles: Color.Blue,
-  bluegrass: Color.Blue,
-  fiddle: Color.Orange,
-  christian: Color.Magenta,
-  cowboy: Color.Orange,
-  club: Color.Purple,
-  exercise: Color.Yellow,
-  hardcore: Color.Red,
-  house: Color.Orange,
-  techno: Color.Magenta,
-  trance: Color.Purple,
-  trap: Color.Orange,
-  swing: Color.Blue,
-  drum: Color.SecondaryText,
-  crunk: Color.Green,
-  glitch: Color.PrimaryText,
-  "hip hop": Color.Purple,
-  rap: Color.Red,
-  karaoke: Color.Red,
-  latin: Color.Orange,
-  soul: Color.SecondaryText,
-  "R&B": Color.Red,
-  reggae: Color.Yellow,
-  african: Color.Yellow,
-  asian: Color.Magenta,
-  australian: Color.Red,
-  cajun: Color.Orange,
-  caribbean: Color.Blue,
-  celtic: Color.Green,
-  european: Color.Magenta,
-  polka: Color.Purple,
-  hawaiian: Color.Yellow,
+interface ExtensionPreferences {
+  audioApplication: string;
+}
+
+export const dummyStation: StationData = {
+  name: "",
+  website: "",
+  stream: "",
+  genres: [],
+  description: "",
+  discontinued: false
 };
 
 interface stationList {
@@ -147,17 +64,23 @@ export async function loadDefaults() {
 
 export async function getAllStations() {
   // Gets a list of all stations and their associated data
-  const stations: { [stationName: string]: { [key: string]: string | string[] } } = {};
+  const stations: StationListObject = {};
   const stationEntries = await LocalStorage.allItems();
   Object.entries(stationEntries).forEach(([key, value]) => {
     if (key.startsWith("station~")) {
       const stationName = key.substring(key.indexOf("~") + 1, key.indexOf("#"));
       const attribute = key.substring(key.indexOf("#") + 1);
-      const trueValue = value.indexOf(",") == -1 ? value : value.split(",");
+      let trueValue = value.toString();
+      trueValue = trueValue.indexOf(",") == -1 ? trueValue : trueValue.split(",");
 
       if (!(stationName in stations)) {
         stations[stationName] = {
           name: stationName,
+          website: "",
+          stream: "",
+          genres: [],
+          description: "",
+          discontinued: false
         };
       }
       stations[stationName][attribute] = trueValue;
@@ -166,39 +89,98 @@ export async function getAllStations() {
   return stations;
 }
 
-export async function playStation(stationName: string, stationURL: string) {
+export async function playStation(stationName: string, stationURL: string): Promise<string | number> {
   // Plays a station's stream URL in Music.app
   const isPlaying = await LocalStorage.getItem("-is-playing");
   const prevTrackID = await LocalStorage.getItem("-current-track-id");
   const prevStationName = await LocalStorage.getItem("-current-station-name");
 
+  const preferences = getPreferenceValues<ExtensionPreferences>();
+  const audioApplication = preferences.audioApplication;
+
   if (isPlaying == "true") {
+    // Pause & remove the current stream before playing a new one
     await pausePlayback();
     await deleteTrack(prevTrackID?.toString(), prevStationName?.toString());
   }
 
-  const streamID = await runAppleScript(`tell application "Music"
-      try
-        set trackIDs to get id of URL tracks
-        open location "${stationURL}"
-        repeat with newID in (get id of URL tracks)
-            if newID is not in trackIDs then
-                set theStream to track id newID
-                set name of theStream to "Raycast: ${stationName}"
-                return contents of newID
-            end if
-        end repeat
+  let streamID = "";
+  if (audioApplication == "music") {
+    streamID = await runAppleScript(`try
+        tell application "Music"
+          tell application "System Events"
+            repeat while "Music" is not in (name of every process whose background only is false)
+              delay 0.5
+            end repeat
+          end tell
+
+          try
+            set trackIDs to get id of URL tracks
+            open location "${stationURL}"
+            repeat with newID in (get id of URL tracks)
+                if newID is not in trackIDs then
+                    set theStream to track id newID
+                    set name of theStream to "Raycast: ${stationName}"
+                    return contents of newID
+                end if
+            end repeat
+          on error
+            open location "${stationURL}"
+            delay 0.5
+            try
+              set name of URL track 1 to "Raycast: ${stationName}"
+              return id of URL track 1
+            on error
+              return ""
+            end try
+          end try
+        end tell
       on error
-        open location "${stationURL}"
-        delay 0.5
-        try
-          set name of URL track 1 to "Raycast: ${stationName}"
-          return id of URL track 1
-        on error
-          return ""
-        end try
-      end try
-    end tell`);
+        return "err:noapp"
+      end try`);
+  } else if (audioApplication == "quicktime") {
+    streamID = await runAppleScript(`try
+        tell application "QuickTime Player"
+          open location "${stationURL}"
+          repeat while visible of window 1 = false
+            delay 0.5
+          end repeat
+          delay 1
+          set visible of window 1 to false
+          return name of document 1
+        end tell
+      on error
+        return "err:noapp"
+      end try`);
+  } else if (audioApplication == "vlc") {
+    streamID = await runAppleScript(`try
+        tell application "VLC"
+          open location "${stationURL}"
+          return path of current item
+        end tell
+      on error
+        return "err:noapp"
+      end try`);
+  } else if (audioApplication == "vox") {
+    streamID = await runAppleScript(`try
+        tell application "Vox"
+          playUrl "${stationURL}"
+          return track 
+        end tell
+      on error
+        return "err:noapp"
+      end try`);
+  }
+
+  if (streamID === "err:noapp") {
+    // Couldn't find target application
+    if (environment.launchType == LaunchType.Background || environment.commandName === "menubar-radio") {
+      await showHUD("Error: Audio Application Not Found!");
+    } else {
+      await showToast({ title: "Error: Audio Application Not Found!", style: Toast.Style.Failure });
+    }
+    return -1;
+  }
 
   await LocalStorage.setItem("-is-playing", "true");
   await LocalStorage.setItem("-current-station-name", stationName);
@@ -215,16 +197,35 @@ export async function playStation(stationName: string, stationURL: string) {
 
 export async function pausePlayback() {
   // Pauses Music.app and clears storage data keys
-
   const stationName = await LocalStorage.getItem("-current-station-name");
   const stationURL = await LocalStorage.getItem("-current-station-url");
+  const trackID = await LocalStorage.getItem("-current-track-id");
 
   await LocalStorage.setItem("-last-station-name", stationName as string);
   await LocalStorage.setItem("-last-station-url", stationURL as string);
 
-  await runAppleScript(`tell application "Music"
-      pause
-    end tell`);
+  const preferences = getPreferenceValues<ExtensionPreferences>();
+  const audioApplication = preferences.audioApplication;
+
+  if (audioApplication == "music") {
+    await runAppleScript(`tell application "Music"
+        pause
+      end tell`);
+  } else if (audioApplication == "quicktime") {
+    await runAppleScript(`tell application "QuickTime Player"
+        try
+          close first document whose name is "${trackID}"
+        end try
+      end tell`);
+  } else if (audioApplication == "vlc") {
+    await runAppleScript(`tell application "VLC"
+        stop
+      end tell`);
+  } else if (audioApplication == "vox") {
+    await runAppleScript(`tell application "Vox"
+        pause
+      end tell`);
+  }
 
   await LocalStorage.setItem("-is-playing", "");
   await LocalStorage.setItem("-current-station-name", "");
@@ -239,6 +240,12 @@ export async function pausePlayback() {
 
 export async function deleteTrack(trackID?: string, trackName?: string) {
   // Deletes a station's track from Music.app
+  const preferences = getPreferenceValues<ExtensionPreferences>();
+  const audioApplication = preferences.audioApplication;
+  if (audioApplication == "quicktime" || audioApplication == "vlc" || audioApplication == "vox") {
+    return;
+  }
+
   if (trackID != undefined && trackID != "") {
     await runAppleScript(`tell application "Music"
             try
@@ -255,28 +262,24 @@ export async function deleteTrack(trackID?: string, trackName?: string) {
   }
 }
 
-export async function deleteStation(stationName: string, stationData: { [key: string]: string | string[] }) {
+export async function getPlayStatus() {
+  return await runAppleScript(`tell application "Music" to return player state`) == "playing"
+}
+
+export async function deleteStation(stationName: string, stationData: StationData) {
   // Deletes a station from storage
   for (const key in stationData) {
+    await LocalStorage.setItem("-temp-station-info", JSON.stringify(stationData));
     await LocalStorage.removeItem(`station~${stationName}#${key}`);
   }
 }
 
-async function modifyStation(
+export async function modifyStation(
   oldName: string,
   newName: string,
-  oldData: { [key: string]: string | string[] },
-  newData: { [key: string]: string | string[] },
-  setStations: React.Dispatch<
-    React.SetStateAction<
-      | {
-          [stationName: string]: {
-            [key: string]: string | string[];
-          };
-        }
-      | undefined
-    >
-  >
+  oldData: StationData,
+  newData: StationData,
+  setStations: React.Dispatch<React.SetStateAction<StationListObject | undefined>>
 ) {
   // Updates the metadata of a station
   if (oldName != "") {
@@ -284,10 +287,11 @@ async function modifyStation(
   }
   for (const key in newData) {
     let value = newData[key];
+    if (value == undefined) continue
     if (Array.isArray(value)) {
       value = value.join(",") + ",";
     }
-    await LocalStorage.setItem(`station~${newName}#${key}`, value);
+    await LocalStorage.setItem(`station~${newName}#${key}`, value.toString());
   }
 
   getAllStations().then((stationList) => {
@@ -297,17 +301,8 @@ async function modifyStation(
 
 export function EditStationForm(props: {
   stationName: string;
-  stationData: { [key: string]: string | string[] };
-  setStations: React.Dispatch<
-    React.SetStateAction<
-      | {
-          [stationName: string]: {
-            [key: string]: string | string[];
-          };
-        }
-      | undefined
-    >
-  >;
+  stationData: StationData;
+  setStations: React.Dispatch<React.SetStateAction<StationListObject | undefined>>;
 }) {
   // Displays a form for updating station metadata
   const { stationName, stationData, setStations } = props;
@@ -370,9 +365,12 @@ export function EditStationForm(props: {
                 values.nameField,
                 stationData,
                 {
+                  name: stationName,
                   website: values.websiteField,
                   stream: values.streamField,
                   genres: values.genresField,
+                  description: "",
+                  discontinued: false
                 },
                 setStations
               );