+++ extensions/internet-radio/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Internet Radio Changelog
 
+## [QuickTime Player, VLC, and VOX] - 2023-02-03
+
+- Add preference setting for target audio application
+- Add support for playback via QuickTime Player, VLC, and VOX
+- Add checks for when target audio application is not installed
+- Add "Delete All Stations" action in "Browse Stations" list
+
 ## [Initial Version] - 2023-01-31